rimeX/requirements.txt
git+https://github.com/ISI-MIP/isimip-client
numpy
xarray
PyYAML
tqdm
